requirements.txt
SQLAlchemy==2.0.25
RPi.GPIO==0.7.0
loguru==0.7.2